backend/requirements.txt
tensorflow-cpu==2.8.0
protobuf==3.19.4
numpy
requests
Pillow
fastapi
python-multipart
uvicorn
gunicorn